+++ packages/shared/src/types/tags.types.ts
@@ -1,12 +1,9 @@
-import {z} from 'zod';
-
-import {makeUuid} from '@shared/lib/utils.shared';
-
-import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
 
 export enum TagType {
+  /** A tag whose lifecycle is managed by the user. */
   User = 'USER',
+  /** A tag whose lifecycle is managed by the system. */
   System = 'SYSTEM',
 }
 
@@ -14,18 +11,6 @@ export enum TagType {
  * Strongly-typed type for a {@link UserTag}'s unique identifier. Prefer this over plain strings.
  */
 export type UserTagId = string & {readonly __brand: 'UserTagIdBrand'};
-
-/**
- * Zod schema for a {@link UserTagId}.
- */
-export const UserTagIdSchema = z.string().uuid();
-
-/**
- * Creates a new random {@link UserTagId}.
- */
-export function makeUserTagId(): UserTagId {
-  return makeUuid<UserTagId>();
-}
 
 /**
  * Unique IDs for {@link SystemTag}s, which are tags whose lifecycle is managed by the system.
@@ -40,11 +25,6 @@ export enum SystemTagId {
   // Done = 'DONE',
 }
 
-/**
- * Zod schema for a {@link SystemTagId}.
- */
-export const SystemTagIdSchema = z.nativeEnum(SystemTagId);
-
 export type TagId = UserTagId | SystemTagId;
 
 /**
@@ -52,7 +32,7 @@ export type TagId = UserTagId | SystemTagId;
  * associated metadata, such as a color and icon. They can also be used as filter criteria for
  * views and searches.
  */
-export interface Tag {
+interface BaseTag {
   readonly tagId: TagId;
   readonly tagType: TagType;
   readonly name: string;
@@ -62,60 +42,15 @@ export interface Tag {
 /**
  * A tag whose lifecycle is managed by the user.
  */
-export interface UserTag extends Tag, BaseStoreItem {
+export interface UserTag extends BaseTag, BaseStoreItem {
   readonly tagType: TagType.User;
   readonly tagId: UserTagId;
 }
 
 /**
- * Zod schema for a {@link UserTag}.
- */
-export const UserTagSchema = z.object({
-  tagId: UserTagIdSchema,
-  name: z.string().min(1).max(255),
-  createdTime: FirestoreTimestampSchema.or(z.date()),
-  lastUpdatedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-/**
  * A tag whose lifecycle is managed by the system.
  */
-export interface SystemTag extends Tag {
+export interface SystemTag extends BaseTag {
   readonly tagType: TagType.System;
   readonly tagId: SystemTagId;
-}
-
-export const SystemTagSchema = z.object({
-  tagId: SystemTagIdSchema,
-  name: z.string().min(1).max(255),
-});
-
-export class Tags {
-  static readonly UNREAD_TAG: Tag = {
-    tagId: SystemTagId.Unread,
-    tagType: TagType.System,
-    name: 'Unread',
-  };
-
-  static readonly STARRED_TAG: Tag = {
-    tagId: SystemTagId.Starred,
-    tagType: TagType.System,
-    name: 'Starred',
-  };
-
-  static readonly TRASHED_TAG: Tag = {
-    tagId: SystemTagId.Trashed,
-    tagType: TagType.System,
-    name: 'Trashed',
-  };
-
-  static makeUserTag(tagInfo: Omit<UserTag, 'tagId' | 'tagType'>): UserTag {
-    return {
-      tagId: makeUserTagId(),
-      tagType: TagType.User,
-      name: tagInfo.name,
-      createdTime: tagInfo.createdTime,
-      lastUpdatedTime: tagInfo.lastUpdatedTime,
-    };
-  }
 }
